+++ packages/sdk/src/abi/wrapper.abi.ts
@@ -141,11 +141,6 @@ export const wrapperAbi = [
   },
   {
     inputs: [],
-    name: "EthFeeTransferFailed",
-    type: "error",
-  },
-  {
-    inputs: [],
     name: "FailedCall",
     type: "error",
   },
@@ -400,31 +395,6 @@ export const wrapperAbi = [
     anonymous: false,
     inputs: [
       {
-        indexed: true,
-        internalType: "address",
-        name: "holder",
-        type: "address",
-      },
-      {
-        indexed: true,
-        internalType: "address",
-        name: "operator",
-        type: "address",
-      },
-      {
-        indexed: false,
-        internalType: "uint48",
-        name: "until",
-        type: "uint48",
-      },
-    ],
-    name: "FinalizeUnwrapOperatorSet",
-    type: "event",
-  },
-  {
-    anonymous: false,
-    inputs: [
-      {
         indexed: false,
         internalType: "uint64",
         name: "version",
@@ -639,55 +609,6 @@ export const wrapperAbi = [
     anonymous: false,
     inputs: [
       {
-        indexed: true,
-        internalType: "euint64",
-        name: "burntAmountHandle",
-        type: "bytes32",
-      },
-      {
-        indexed: false,
-        internalType: "bool",
-        name: "finalizeSuccess",
-        type: "bool",
-      },
-      {
-        indexed: false,
-        internalType: "bool",
-        name: "feeTransferSuccess",
-        type: "bool",
-      },
-      {
-        indexed: false,
-        internalType: "uint64",
-        name: "burnAmount",
-        type: "uint64",
-      },
-      {
-        indexed: false,
-        internalType: "uint256",
-        name: "unwrapAmount",
-        type: "uint256",
-      },
-      {
-        indexed: false,
-        internalType: "uint256",
-        name: "feeAmount",
-        type: "uint256",
-      },
-      {
-        indexed: true,
-        internalType: "uint256",
-        name: "nextTxId",
-        type: "uint256",
-      },
-    ],
-    name: "UnwrappedFinalized",
-    type: "event",
-  },
-  {
-    anonymous: false,
-    inputs: [
-      {
         indexed: false,
         internalType: "bool",
         name: "returnVal",
@@ -750,33 +671,15 @@ export const wrapperAbi = [
     anonymous: false,
     inputs: [
       {
-        indexed: false,
-        internalType: "uint64",
-        name: "mintAmount",
-        type: "uint64",
+        indexed: true,
+        internalType: "address",
+        name: "to",
+        type: "address",
       },
       {
         indexed: false,
         internalType: "uint256",
         name: "amountIn",
-        type: "uint256",
-      },
-      {
-        indexed: false,
-        internalType: "uint256",
-        name: "feeAmount",
-        type: "uint256",
-      },
-      {
-        indexed: true,
-        internalType: "address",
-        name: "to_",
-        type: "address",
-      },
-      {
-        indexed: true,
-        internalType: "uint256",
-        name: "mintTxId",
         type: "uint256",
       },
     ],
@@ -801,19 +704,6 @@ export const wrapperAbi = [
     name: "acceptOwnership",
     outputs: [],
     stateMutability: "nonpayable",
-    type: "function",
-  },
-  {
-    inputs: [],
-    name: "adminProvider",
-    outputs: [
-      {
-        internalType: "contract AdminProvider",
-        name: "",
-        type: "address",
-      },
-    ],
-    stateMutability: "view",
     type: "function",
   },
   {
@@ -1140,19 +1030,6 @@ export const wrapperAbi = [
     type: "function",
   },
   {
-    inputs: [],
-    name: "deploymentCoordinator",
-    outputs: [
-      {
-        internalType: "contract IDeploymentCoordinator",
-        name: "",
-        type: "address",
-      },
-    ],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
     inputs: [
       {
         internalType: "euint64",
@@ -1225,39 +1102,10 @@ export const wrapperAbi = [
         name: "owner_",
         type: "address",
       },
-      {
-        internalType: "contract IDeploymentCoordinator",
-        name: "deploymentCoordinator_",
-        type: "address",
-      },
     ],
     name: "initialize",
     outputs: [],
     stateMutability: "nonpayable",
-    type: "function",
-  },
-  {
-    inputs: [
-      {
-        internalType: "address",
-        name: "holder",
-        type: "address",
-      },
-      {
-        internalType: "address",
-        name: "operator",
-        type: "address",
-      },
-    ],
-    name: "isFinalizeUnwrapOperator",
-    outputs: [
-      {
-        internalType: "bool",
-        name: "",
-        type: "bool",
-      },
-    ],
-    stateMutability: "view",
     type: "function",
   },
   {
@@ -1453,24 +1301,6 @@ export const wrapperAbi = [
       },
     ],
     stateMutability: "view",
-    type: "function",
-  },
-  {
-    inputs: [
-      {
-        internalType: "address",
-        name: "operator",
-        type: "address",
-      },
-      {
-        internalType: "uint48",
-        name: "until",
-        type: "uint48",
-      },
-    ],
-    name: "setFinalizeUnwrapOperator",
-    outputs: [],
-    stateMutability: "nonpayable",
     type: "function",
   },
   {
@@ -1805,24 +1635,6 @@ export const wrapperAbi = [
     name: "wrap",
     outputs: [],
     stateMutability: "nonpayable",
-    type: "function",
-  },
-  {
-    inputs: [
-      {
-        internalType: "address",
-        name: "to",
-        type: "address",
-      },
-      {
-        internalType: "uint256",
-        name: "amount",
-        type: "uint256",
-      },
-    ],
-    name: "wrapETH",
-    outputs: [],
-    stateMutability: "payable",
     type: "function",
   },
 ] as const;