+++ packages/sdk/src/viem/index.ts
@@ -8,18 +8,14 @@
 export { ViemSigner, type ViemSignerConfig } from "./viem-signer";
 export {
   readConfidentialBalanceOfContract,
-  readWrapperForTokenContract,
   readUnderlyingTokenContract,
-  readWrapperExistsContract,
   readSupportsInterfaceContract,
   writeConfidentialTransferContract,
-  writeConfidentialBatchTransferContract,
   writeUnwrapContract,
   writeUnwrapFromBalanceContract,
   writeFinalizeUnwrapContract,
   writeSetOperatorContract,
   writeWrapContract,
-  writeWrapETHContract,
   readTokenPairsContract,
   readTokenPairsLengthContract,
   readTokenPairsSliceContract,